roster input added and returns input to screen
